+++ src/main/kotlin/solver/legacy/ProblemWriterNoDummy.kt
@@ -12,7 +12,7 @@ class ProblemWriterNoDummy : ProblemWriter {
         distanceMatrix: Array<DoubleArray>,
         cluster: Cluster,
         path: String,
-        costLimit: Int
+        costLimit: Double
     ) {
         val nodeList: MutableList<Node> = mutableListOf()
         val clusterDistanceMatrix = createDistanceMatrix(cluster, distanceMatrix, nodeList)
